Move Triton ensemble code out of init file
